requirements.txt
jupyter
jupytext
numpy
opencv
pillow
scipy
